+++ lib/help.rb
@@ -103,6 +103,7 @@ class AyaDN
 		    commands << "ayadn repost 18527205\n"
 		    commands << "ayadn unrepost 18527205\n"
 		    commands << "ayadn delete 12345678\n"
+		    commands << "ayadn delete-message 12345 23456789\n"
 		    commands << "ayadn list muted\n"
 		    commands << "ayadn list followings @ericd\n"
 		    commands << "ayadn list followers @ericd\n"