+++ CHANGELOG.md
@@ -3,6 +3,10 @@ AYADN
 
 AYADN is a learning exercise: while it's fully functional, it's still very alpha (remember I'm a beginner with Ruby).  
 
+## 2014-01-04 v0.5.5
+- delete messages
+- users database
+
 ## 2013-12-30 v0.5.4
 - custom posts count by stream